рефакторинг коду (дані одного користувача не доступні іншому)

+++ src/boards/repository.ts
@@ -12,7 +12,7 @@ export class BoardsRepository{
     async create(data:any){
         const result = await this.pool.query(
             "INSERT INTO boards (name,user_id) VALUES ($1,$2) RETURNING *",
-            [data.name,data.user_id]
+            [data.name, data.userId]
         );
 
         return result.rows[0];
@@ -20,8 +20,8 @@ export class BoardsRepository{
 
     async delete(data:any){
         const result = await this.pool.query(
-            "DELETE FROM boards WHERE id = $1 RETURNING *",
-            [data.id]
+            "DELETE FROM boards WHERE id = $1 AND user_id = $2 RETURNING *",
+            [data.id, data.userId]
         );
 
         if(result.rows[0]) return result.rows[0];
@@ -30,26 +30,26 @@ export class BoardsRepository{
 
     async updateName(data:any){
         const result = await this.pool.query(
-            "UPDATE boards SET name = $1 WHERE id = $2 RETURNING *",
-            [data.name,data.id]
+            "UPDATE boards SET name = $1 WHERE id = $2 AND user_id = $3 RETURNING *",
+            [data.name,data.id, data.userId]
         );
 
         if(result.rows[0]) return result.rows[0];
         else throw new AppError(404,"Board not found");
     }
-
+/*
     async readAll(){
         const result = await this.pool.query(
             "SELECT * FROM boards"
         );
 
         return result.rows;
-    }
+    }*/
 
     async readByUser(data:any){
         const result = await this.pool.query(
             "SELECT * FROM boards WHERE user_id = $1",
-            [data.user_id]
+            [data.userId]
         );
 
         return result.rows;
