+++ src/boards/router.ts
@@ -5,26 +5,45 @@ import { AppError } from "../errorHandler/service";
 export const boardRoutes = Router();
 const service = new BoardsService();
 
-boardRoutes.get("/GetAllBoards", async(req,res)=>{
-    const result = await service.GetAllBoards();
-    return res.status(200).json(result);
-});
+
 
 boardRoutes.post("/AddBoard", async(req,res)=>{
-    const {name,user_id} = req.body;
+    const userId = req.user!.userId;
+    const {name} = req.body;
 
     if(!name) throw new AppError(401,"Name required");
-    if(!user_id) throw new AppError(401,"user_id required");
 
-    const result = await service.AddBoard({name,user_id});
+    const result = await service.AddBoard(userId, {name});
     return res.status(200).json(result);
 });
 
 boardRoutes.delete("/DeleteBoard", async(req,res)=>{
+    const userId = req.user!.userId;
     const {id} = req.body;
 
-    const result = await service.DeleteBoard({id});
+    const result = await service.DeleteBoard(userId, {id});
     return res.status(200).json(result);
 });
+
+boardRoutes.get("/GetBoardsByUser", async(req,res)=>{
+    const userId = req.user!.userId;
+
+    const result = await service.GetBoardsByUser(userId);
+    return res.status(200).json(result);
+});
+
+boardRoutes.put("/UpdateBoardName", async(req,res)=>{
+    const userId = req.user!.userId;
+    const {name} = req.body;
+
+    const result = await service.UpdateBoardName(userId, {name});
+    return res.status(200).json(result);
+});
+
+/*
+boardRoutes.get("/GetAllBoards", async(req,res)=>{
+    const result = await service.GetAllBoards();
+    return res.status(200).json(result);
+});*/
 
 export default boardRoutes
