+++ src/tasks/service.ts
@@ -25,12 +25,12 @@ export class TasksService{
         this.repository = new TasksRepository();
     }
 
-    async AddTask(payload: AddTaskDTO){
+    async AddTask(userId: number, payload: AddTaskDTO){
         //const userId = req.user?.userId;
         const {title, description} = payload;
         
         try {
-            return await this.repository.create({title, description});
+            return await this.repository.create({userId, title, description});
         } catch(e){
             if (e instanceof AppError) throw e;
             throw new AppError(500, "Error");
@@ -38,20 +38,20 @@ export class TasksService{
         
     }
 
-    async GetAllTasks(){
+    /*async GetAllTasks(){
         try {
             return await this.repository.readAll();
         } catch(e){
             if (e instanceof AppError) throw e;
             throw new AppError(500, "Error");
         }
-    }
+    }*/
 
-    async GetTaskByTitle(payload: GetTaskByTitleDTO){
+    async GetTaskByTitle(userId: number, payload: GetTaskByTitleDTO){
         const {title} = payload;
         
         try {
-            return await this.repository.readTitle({title});
+            return await this.repository.readTitle({userId, title});
         } catch(e){
             if (e instanceof AppError) throw e;
             throw new AppError(500, "Error");
@@ -59,11 +59,11 @@ export class TasksService{
         
     }
 
-    async GetTaskByPriority(payload: GetTaskByPriorityDTO){
+    async GetTaskByPriority(userId: number, payload: GetTaskByPriorityDTO){
         const {priority} = payload;
         
         try {
-            return await this.repository.readPriority({priority});
+            return await this.repository.readPriority({userId, priority});
         } catch(e){
             if (e instanceof AppError) throw e;
             throw new AppError(500, "Error");
@@ -71,11 +71,11 @@ export class TasksService{
         
     }
 
-    async UpdateTaskTitle(payload: UpdateTaskTitleDTO){
+    async UpdateTaskTitle(userId: number, payload: UpdateTaskTitleDTO){
         const {task_id, title} = payload;
         
         try {
-            return await this.repository.updateTitle({task_id, title});
+            return await this.repository.updateTitle({userId, task_id, title});
         } catch(e){
             if (e instanceof AppError) throw e;
             throw new AppError(500, "Error");
@@ -83,11 +83,11 @@ export class TasksService{
         
     }
 
-    async UpdateTaskDescription(payload: UpdateTaskDescriptionDTO){
+    async UpdateTaskDescription(userId: number, payload: UpdateTaskDescriptionDTO){
         const {task_id, description} = payload;
         
         try {
-            return await this.repository.updateDescription({task_id, description});
+            return await this.repository.updateDescription({userId, task_id, description});
         } catch(e){
             if (e instanceof AppError) throw e;
             throw new AppError(500, "Error");
@@ -95,11 +95,11 @@ export class TasksService{
         
     }
 
-    async UpdateTaskDedline(payload: UpdateTaskDedlineDTO){
+    async UpdateTaskDedline(userId: number, payload: UpdateTaskDedlineDTO){
         const {task_id, due_date} = payload;
         
         try {
-            return await this.repository.updateDeadline({task_id, due_date});
+            return await this.repository.updateDeadline({userId, task_id, due_date});
         } catch(e){
             if (e instanceof AppError) throw e;
             throw new AppError(500, "Error");
@@ -107,11 +107,11 @@ export class TasksService{
         
     }
 
-    async UpdateTaskPriority(payload: UpdateTaskPriorityDTO){
+    async UpdateTaskPriority(userId: number, payload: UpdateTaskPriorityDTO){
         const {task_id, priority} = payload;
         
         try {
-            return await this.repository.updatePriority({task_id, priority});
+            return await this.repository.updatePriority({userId, task_id, priority});
         } catch(e){
             if (e instanceof AppError) throw e;
             throw new AppError(500, "Error");
@@ -119,11 +119,11 @@ export class TasksService{
         
     }
 
-    async deleteTask(payload: DeleteTaskDTO){
+    async deleteTask(userId: number, payload: DeleteTaskDTO){
         const {task_id} = payload;
         
         try {
-            return await this.repository.delete({task_id});
+            return await this.repository.delete({userId, task_id});
         } catch(e){
             if (e instanceof AppError) throw e;
             throw new AppError(500, "Error");
